packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aaa-core"
version = "0.1.0"
description = "Core modules for Access Ability Arm (config, hardware, workers)"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [
    {name = "Access Ability Arm Team"}
]
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
realsense = [
    "pyrealsense2>=2.56.0",
]

[tool.setuptools.packages.find]
where = ["src"]
